I've resolved the circular import issue between `app_factory` and `api_system`.
To do this, I moved the `reschedule_unified_backup_jobs` function from
`app_factory.py` to `utils.py`. This broke the circular dependency
that was happening because `routes.api_system.py` tried to import
this function while `app_factory.py` was also importing
modules from `routes.api_system.py`.

Here's what I changed:
- I defined `reschedule_unified_backup_jobs` in `utils.py`.
- I updated `utils.py` with the necessary imports for this function
  (scheduler tasks) and made sure it uses the passed `app_instance`
  for logger and scheduler access.
- I removed the function definition from `app_factory.py`.
- I updated `routes.api_system.py` to import the function from
  `utils.py`.

This change should fix the `ImportError` you were seeing when the application started.
